feat: pdf-view -> renamed to pdf
