+++ apps/demo-angular/src/app-routing.module.ts
@@ -34,7 +34,7 @@ const routes: Routes = [
 	{ path: 'iqkeyboardmanager', loadChildren: () => import('./plugin-demos/iqkeyboardmanager.module').then((m) => m.IqkeyboardmanagerModule) },
 	{ path: 'local-notifications', loadChildren: () => import('./plugin-demos/local-notifications.module').then((m) => m.LocalNotificationsModule) },
 	{ path: 'localize', loadChildren: () => import('./plugin-demos/localize.module').then((m) => m.LocalizeModule) },
-	{ path: 'pdf-view', loadChildren: () => import('./plugin-demos/pdf-view.module').then(m => m.PdfViewModule) },
+	{ path: 'pdf', loadChildren: () => import('./plugin-demos/pdf-view.module').then(m => m.PdfViewModule) },
 	{ path: 'picker', loadChildren: () => import('./plugin-demos/picker.module').then((m) => m.PickerModule) },
 	{ path: 'shared-notification-delegate', loadChildren: () => import('./plugin-demos/shared-notification-delegate.module').then((m) => m.SharedNotificationDelegateModule) },
 	{ path: 'social-share', loadChildren: () => import('./plugin-demos/social-share.module').then((m) => m.SocialShareModule) },
