--- apps/demo-angular/src/plugin-demos/pdf-view.component.html
@@ -1,8 +1,0 @@
-<ActionBar title="pdf-view" class="action-bar"> </ActionBar>
-<StackLayout class="p-20">
-  <ScrollView class="h-full">
-    <StackLayout>
-      <Button text="Test nativescript-pdf-view" (tap)="demoShared.testIt()" class="btn btn-primary"></Button>
-    </StackLayout>
-  </ScrollView>
-</StackLayout>
