+++ apps/demo-angular/src/plugin-demos/pdf.component.html
@@ -1,0 +1,8 @@
+<ActionBar title="pdf" class="action-bar"> </ActionBar>
+<StackLayout class="p-20">
+	<ScrollView class="h-full">
+		<StackLayout>
+			<Button text="Test nativescript-pdf" (tap)="demoShared.testIt()" class="btn btn-primary"></Button>
+		</StackLayout>
+	</ScrollView>
+</StackLayout>
