+++ apps/demo-angular/src/plugin-demos/pdf.module.ts
@@ -1,0 +1,10 @@
+import { NgModule, NO_ERRORS_SCHEMA } from '@angular/core';
+import { NativeScriptCommonModule, NativeScriptRouterModule } from '@nativescript/angular';
+import { PdfComponent } from './pdf.component';
+
+@NgModule({
+	imports: [NativeScriptCommonModule, NativeScriptRouterModule.forChild([{ path: '', component: PdfComponent }])],
+	declarations: [PdfComponent],
+	schemas: [NO_ERRORS_SCHEMA],
+})
+export class PdfModule {}
