+++ apps/demo/package.json
@@ -32,7 +32,7 @@
 		"@nativescript/iqkeyboardmanager": "file:../../packages/iqkeyboardmanager",
 		"@nativescript/local-notifications": "file:../../packages/local-notifications",
 		"@nativescript/localize": "file:../../dist/packages/localize",
-		"@nativescript/pdf-view": "file:../../packages/pdf-view",
+		"@nativescript/pdf": "file:../../packages/pdf",
 		"@nativescript/picker": "file:../../packages/picker",
 		"@nativescript/shared-notification-delegate": "file:../../packages/shared-notification-delegate",
 		"@nativescript/social-share": "file:../../packages/social-share",
