--- packages/pdf-view/project.json
@@ -1,72 +1,0 @@
-{
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "sourceRoot": "packages/pdf-view",
-  "targets": {
-    "build": {
-      "executor": "@nrwl/js:tsc",
-      "options": {
-        "outputPath": "dist/packages/pdf-view",
-        "tsConfig": "packages/pdf-view/tsconfig.json",
-        "packageJson": "packages/pdf-view/package.json",
-        "main": "packages/pdf-view/index.d.ts",
-        "assets": [
-          "packages/pdf-view/*.md",
-          "packages/pdf-view/index.d.ts",
-          "LICENSE",
-          {
-            "glob": "**/*",
-            "input": "packages/pdf-view/platforms/",
-            "output": "./platforms/"
-          }
-        ],
-        "dependsOn": [
-          {
-            "target": "build.all",
-            "projects": "dependencies"
-          }
-        ]
-      }
-    },
-    "build.all": {
-      "executor": "@nrwl/workspace:run-commands",
-      "options": {
-        "commands": [
-          "node tools/scripts/build-finish.ts pdf-view"
-        ],
-        "parallel": false
-      },
-      "outputs": [
-        "dist/packages/pdf-view"
-      ],
-      "dependsOn": [
-        {
-          "target": "build.all",
-          "projects": "dependencies"
-        },
-        {
-          "target": "build",
-          "projects": "self"
-        }
-      ]
-    },
-    "focus": {
-      "executor": "@nrwl/workspace:run-commands",
-      "options": {
-        "commands": [
-          "nx g @nativescript/plugin-tools:focus-packages pdf-view"
-        ],
-        "parallel": false
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "options": {
-        "lintFilePatterns": [
-          "packages/pdf-view/**/*.ts"
-        ]
-      }
-    }
-  },
-  "tags": []
-}
