--- packages/pdf-view/references.d.ts
@@ -1,1 +1,0 @@
-/// <reference path="../../references.d.ts" />
