--- packages/pdf-view/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "rootDir": "."
-  },
-  "exclude": ["**/*.spec.ts", "**/*.test.ts", "angular"],
-  "include": [
-    "**/*.ts",
-    "references.d.ts",
-    "../../node_modules/@nativescript/core/global-types.d.ts"
-  ]
-}
