+++ packages/pdf/README.md
@@ -15,7 +15,7 @@ My aim is to keep the features consistent across iOS and Android.
 ## Installation
 
 ```
-tns plugin add nativescript-pdf-view
+tns plugin add @nativescript/pdf
 ```
 
 ## Usage
@@ -25,7 +25,7 @@ tns plugin add nativescript-pdf-view
 ```xml
 <Page
   xmlns="http://schemas.nativescript.org/tns.xsd"
-  xmlns:pdf="nativescript-pdf-view"
+  xmlns:pdf="@nativescript/pdf"
   loaded="pageLoaded">
   <pdf:PDFView src="{{ pdfUrl }}" load="{{ onLoad }}" />
 </Page>
@@ -34,7 +34,7 @@ tns plugin add nativescript-pdf-view
 ### Angular NativeScript
 
 ```ts
-import { PDFView } from 'nativescript-pdf-view';
+import { PDFView } from '@nativescript/pdf';
 import { registerElement } from 'nativescript-angular';
 registerElement('PDFView', () => PDFView);
 ```
