+++ packages/pdf/platforms/android/include.gradle
