+++ packages/pdf/references.d.ts
@@ -1,0 +1,1 @@
+/// <reference path="../../references.d.ts" />
