+++ tools/demo/pdf/index.ts
@@ -1,8 +1,7 @@
 import { DemoSharedBase } from '../utils';
 
 export class DemoSharedNativescriptPdfView extends DemoSharedBase {
-
-  testIt() {
-    console.log('test pdf-view!');
-  }
+	testIt() {
+		console.log('test pdf!');
+	}
 }
